refactor: rename ValueUtil.mts to reflect its scope
Especially now that the folder already provides some context
