+++ rt/src/Asserts.mts
@@ -1,6 +1,6 @@
 import { Thread, Capability } from './Thread.mjs';
 
-import { isListFlagSet, isTupleFlagSet } from './base/ValuesUtil.mjs';
+import { isListFlagSet, isTupleFlagSet } from './base/rawUtil.mjs';
 import * as proc from './base/ProcessID.mjs';
 const ProcessID = proc.ProcessID;
 import { AbstractLevel } from './AbstractLevel.mjs';
