+++ rt/src/MailboxProcessor.mts
@@ -1,5 +1,5 @@
 import { assertIsHandler, assertIsNTuple, assertIsFunction } from "./Asserts.mjs";
-import { mkTuple } from "./base/ValuesUtil.mjs";
+import { mkTuple } from "./base/rawUtil.mjs";
 import { SchedulerInterface } from "./SchedulerInterface.mjs";
 import { unitLVal } from "./base/unitLVal.mjs";
 import { RuntimeInterface } from "./RuntimeInterface.mjs";
