+++ rt/src/base/rawUtil.mts
