+++ rt/src/builtins/listToTuple.mts
@@ -2,7 +2,7 @@
 import { UserRuntimeZero, Constructor, mkBase } from './UserRuntimeZero.mjs'
 import { LVal } from '../base/LVal.mjs';
 import { assertIsList } from '../Asserts.mjs'
-import { mkTuple } from '../base/ValuesUtil.mjs'
+import { mkTuple } from '../base/rawUtil.mjs'
 import { lub } from '../Level.mjs';
 
 export function BuiltinListToTuple<TBase extends Constructor<UserRuntimeZero>>(Base: TBase) {
